refactor: refactored devtools browser extension scripts to improve ports management and service worker lifetime
